% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_batch_get_application_revisions}
\alias{codedeploy_batch_get_application_revisions}
\title{Gets information about one or more application revisions}
\usage{
codedeploy_batch_get_application_revisions(applicationName, revisions)
}
\arguments{
\item{applicationName}{[required] The name of an AWS CodeDeploy application about which to get revision
information.}

\item{revisions}{[required] Information to get about the application revisions, including type and
location.}
}
\description{
Gets information about one or more application revisions.
}
\section{Request syntax}{
\preformatted{codedeploy$batch_get_application_revisions(
  applicationName = "string",
  revisions = list(
    list(
      revisionType = "S3"|"GitHub"|"String"|"AppSpecContent",
      s3Location = list(
        bucket = "string",
        key = "string",
        bundleType = "tar"|"tgz"|"zip"|"YAML"|"JSON",
        version = "string",
        eTag = "string"
      ),
      gitHubLocation = list(
        repository = "string",
        commitId = "string"
      ),
      string = list(
        content = "string",
        sha256 = "string"
      ),
      appSpecContent = list(
        content = "string",
        sha256 = "string"
      )
    )
  )
)
}
}

\keyword{internal}
