% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_batch_get_on_premises_instances}
\alias{codedeploy_batch_get_on_premises_instances}
\title{Gets information about one or more on-premises instances}
\usage{
codedeploy_batch_get_on_premises_instances(instanceNames)
}
\arguments{
\item{instanceNames}{[required] The names of the on-premises instances about which to get information.}
}
\description{
Gets information about one or more on-premises instances.
}
\section{Request syntax}{
\preformatted{codedeploy$batch_get_on_premises_instances(
  instanceNames = list(
    "string"
  )
)
}
}

\keyword{internal}
