% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_get_application}
\alias{codedeploy_get_application}
\title{Gets information about an application}
\usage{
codedeploy_get_application(applicationName)
}
\arguments{
\item{applicationName}{[required] The name of an AWS CodeDeploy application associated with the IAM user
or AWS account.}
}
\description{
Gets information about an application.
}
\section{Request syntax}{
\preformatted{codedeploy$get_application(
  applicationName = "string"
)
}
}

\keyword{internal}
