% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_list_deployments}
\alias{codedeploy_list_deployments}
\title{Lists the deployments in a deployment group for an application
registered with the IAM user or AWS account}
\usage{
codedeploy_list_deployments(applicationName, deploymentGroupName,
  includeOnlyStatuses, createTimeRange, nextToken)
}
\arguments{
\item{applicationName}{The name of an AWS CodeDeploy application associated with the IAM user
or AWS account.}

\item{deploymentGroupName}{The name of a deployment group for the specified application.}

\item{includeOnlyStatuses}{A subset of deployments to list by status:
\itemize{
\item Created: Include created deployments in the resulting list.
\item Queued: Include queued deployments in the resulting list.
\item In Progress: Include in-progress deployments in the resulting list.
\item Succeeded: Include successful deployments in the resulting list.
\item Failed: Include failed deployments in the resulting list.
\item Stopped: Include stopped deployments in the resulting list.
}}

\item{createTimeRange}{A time range (start and end) for returning a subset of the list of
deployments.}

\item{nextToken}{An identifier returned from the previous list deployments call. It can
be used to return the next set of deployments in the list.}
}
\description{
Lists the deployments in a deployment group for an application
registered with the IAM user or AWS account.
}
\section{Request syntax}{
\preformatted{codedeploy$list_deployments(
  applicationName = "string",
  deploymentGroupName = "string",
  includeOnlyStatuses = list(
    "Created"|"Queued"|"InProgress"|"Succeeded"|"Failed"|"Stopped"|"Ready"
  ),
  createTimeRange = list(
    start = as.POSIXct(
      "2015-01-01"
    ),
    end = as.POSIXct(
      "2015-01-01"
    )
  ),
  nextToken = "string"
)
}
}

\keyword{internal}
